Using shared metadata for PhoneNumberKit instance
Since metadata is read-only data, it's safe to use it directly from
any thread.

This is not the case for regex manager (see #193). So the idea of this
change is to have smallest change possible to allow having multiple
`PhoneNumberKit` instances which will have metadata shared and will
have regex manager per instance.

Now on the client side it should be possible to create thread-safety
wrappers.